refactor: DeleteExpenseCategoryModal 리팩토링
