+++ src/app/home/setting/expenseCategory/_components/DeleteExpenseCategoryModal.type.ts
@@ -1,0 +1,11 @@
+import { ExpenseCategoryAction } from "../page";
+
+export interface DeleteExpenseCategoryModalProps {
+    currentExpenseCategoryId: number;
+    expenseCategoriesDispatch: React.Dispatch<ExpenseCategoryAction>;
+}
+
+export interface UseDeleteExpenseCategoryModalProps {
+    currentExpenseCategoryId: number;
+    expenseCategoriesDispatch: React.Dispatch<ExpenseCategoryAction>;
+}
